refactor: Abstract view event logic out of MviViewModel.kt
