+++ mvi/src/main/java/dev/jimmymorales/mviandroidtemplate/mvi/events/ConsumableEvent.kt
@@ -1,4 +1,4 @@
-package dev.jimmymorales.mviandroidtemplate.mvi
+package dev.jimmymorales.mviandroidtemplate.mvi.events
 
 import java.util.concurrent.atomic.AtomicBoolean
